+++ pkg/githubclient/client_repository.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"net/http"
+	"regexp"
 
 	"github.com/giantswarm/microerror"
 	"github.com/google/go-github/v44/github"
@@ -247,13 +248,32 @@ func (c *Client) getGithubChecks(ctx context.Context, repository *github.Reposit
 	var checks []string
 	for _, combinedStatus := range allCombinedStatus {
 		for _, status := range combinedStatus.Statuses {
-			checks = append(checks, *status.Context)
+			if filterCheck(status.GetContext()) {
+				checks = append(checks, status.GetContext())
+			}
 		}
 	}
 
 	c.logger.Debugf("found %d commit statuses for ref %q:\n%v", len(checks), ref, checks)
+	for id, check := range checks {
+		c.logger.Debugf(" - checks[%d] = %q", id, check)
+	}
 
 	return checks, nil
+}
+
+// Returns false if check should be ignored
+func filterCheck(checkName string) bool {
+	// List keywords for checks that should be ignored
+	var ignoredCheckKeywords = []string{"aliyun"}
+
+	for _, ignoredCheckKeyword := range ignoredCheckKeywords {
+		matched, _ := regexp.MatchString(ignoredCheckKeyword, checkName)
+		if matched {
+			return false
+		}
+	}
+	return true
 }
 
 // Retrieve list of tags